Made actual assemble function Made passing input file path in arguments Made label creation Fixed #1

+++ main.zig
@@ -1,54 +1,77 @@
 const std = @import("std");
 const print = @import("std").debug.print;
 const asmbl = @import("assembler.zig");
-const instr = @import("instructions.zig");
+const label = @import("labels.zig");
 
 const allocator = std.heap.page_allocator;
 
-const assemblyError = error{unknownPattern};
-
 pub fn main() !void {
     var dir = std.fs.cwd();
-    dir.deleteFile("a.c8") catch {};
+    var args = std.process.args();
+    _ = args.skip();
 
-    var out = try dir.createFile("a.c8", .{});
+    //Arguments parsing
+    var source: []u8 = undefined;
+    if (args.next(allocator)) |src| {
+        source = try src;
+    } else {
+        print("You need to specify the file to assemble in the program arguments.\n", .{});
+        return;
+    }
+    defer allocator.free(source);
 
-    const testProgram = @embedFile("test-programs/blink.asm");
+    //Opening the file
+    print("Opening source file '{}'...\n", .{source});
 
-    var bufLen = std.mem.replacementSize(u8, testProgram, "\x0D\x0A", "\n");
+    var src_file: ?std.fs.File = undefined;
 
-    var buffer = try allocator.alloc(u8, bufLen);
-    defer allocator.free(buffer);
+    if (std.fs.path.isAbsolute(source)) {
+        src_file = std.fs.openFileAbsolute(source, .{}) catch null;
+    } else {
+        src_file = std.fs.Dir.openFile(dir, source, .{}) catch null;
+    }
 
-    _ = std.mem.replace(u8, testProgram, "\x0D\x0A", "\n", buffer);
+    if (src_file) |source_file| {
+        //File is open, do the thing
+        defer source_file.close();
 
-    var lines = std.mem.tokenize(buffer[0..], "\n");
+        print("Creating output file '{}'...\n", .{"a.c8"});
 
-    var line = lines.next();
+        dir.deleteFile("a.c8") catch {};
+        var out = try dir.createFile("a.c8", .{});
 
-    while (line != null) : (line = lines.next()) {
-        print("Line: '{}'.\n", .{line.?});
-        var tokens = std.mem.tokenize(line.?, " ");
+        print("Formatting source...\n", .{});
 
-        const pattern = try asmbl.getPattern(tokens, allocator);
-        defer allocator.free(pattern);
+        var loaded_program: []u8 = undefined;
+        {
+            var loaded = try source_file.readAllAlloc(allocator, (try source_file.stat()).size, std.math.maxInt(usize));
+            defer allocator.free(loaded);
 
-        print("Pattern: '{}'\n", .{pattern});
+            //Replace CRLF with LF
+            var buf_len = std.mem.replacementSize(u8, loaded, "\x0D\n", "\n");
+            loaded_program = try allocator.alloc(u8, buf_len);
 
-        var patId = instr.matchPattern(pattern);
+            _ = std.mem.replace(u8, loaded, "\x0D\n", "\n", loaded_program);
+        }
+        defer allocator.free(loaded_program);
 
-        print("Pattern id: '{}'\n", .{patId.?});
+        //Separate lines
+        var lines = std.mem.tokenize(loaded_program[0..], "\n");
 
-        if (patId == null)
-            return assemblyError.unknownPattern;
+        try label.initLabels(allocator);
+        defer label.deinit();
 
-        const args = try asmbl.extractArgs(tokens, instr.ops[patId.?], allocator);
-        defer allocator.free(args);
+        print("Assembly...\n", .{});
 
-        const bin: u16 = try asmbl.fmtOpCode(patId.?, args);
+        var binary = try asmbl.assembleLines(lines, allocator);
+        defer allocator.free(binary);
 
-        _ = try out.write(@bitCast([2]u8, bin)[0..]);
+        _ = try out.write(binary);
 
-        print("0x{x:0<4}\n", .{bin});
+        print("Done!\n", .{});
+
+    } else {
+        print("Source file could not be opened.\n", .{});
+        return;
     }
 }
